requirements.txt
flask==2.3.3
flask-socketio==5.3.4
qrcode==7.4.2
pillow==10.0.1
gunicorn==21.2.0
gevent==23.9.1
gevent-websocket==0.10.1
python-engineio==4.7.1
python-socketio==5.9.0
